requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
bcrypt==3.2.0
beautifulsoup4==4.9.3
branca==0.4.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-bootstrap4==3.0.1
django-braces==1.14.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==8.1.1
folium==0.12.1
gunicorn==20.1.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
misaka==2.1.1
mysqlclient==2.0.3
numpy==1.21.1
phonenumbers==8.12.28
Pillow==8.2.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.26.0
six==1.15.0
soupsieve==2.2.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.2.0
wincertstore==0.2
